pyproject.toml
[project]
name = "multi-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Amna Ashraf", email = "aamnaashraf501@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]

[project.scripts]
multi-agents = "multi_agents:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
